Fix HDF5resource classes to not cast down longs in close()

+++ uk.ac.gda.nexus/src/gda/data/nexus/hdf5/HDF5AttributeResource.java
@@ -41,7 +41,7 @@ public class HDF5AttributeResource extends HDF5BaseResource {
 			return;
 		}
 		try {
-			H5.H5Aclose((int)resource);
+			H5.H5Aclose(resource);
 		} catch (HDF5LibraryException e) {
 			logger.error("Could not close HDF5 attribute", e);
 		}
